scraping/lagaceta/requirements.txt
playwright==1.41.0
asyncpg==0.29.0
python-dotenv==1.0.0
openai==1.58.1
